DemoPlots

Documentation for DemoPlots.

DemoPlots.plot_chain

DemoPlots.plot_demography

DemoPlots.plot_hist

DemoPlots.plot_input

DemoPlots.plot_remnbps

DemoPlots.plot_residuals_sim

DemoPlots.plot_residuals_th

DemoPlots.xy

DemoPlots.plot_chain — Method

plot_chain(fit::DemoInfer.FitResult, n::Int, ax; kwargs...)

Plot the chain stored in fit for parameter n-th.

ax is the pyplot ax where to plot the chain. Optional arguments are passed to plot from pyplot.

DemoPlots.plot_demography — Method

plot_demography(para::Vector{T}, stderrors::Vector{T} ax; kwargs...)
plot_demography(fit::DemoInfer.FitResult, ax; kwargs...)

Plot the demographic profile encoded in the parameters inferred by the fit.

ax is the pyplot ax where to plot the demographic profile.

Arguments

max_t = 5e6: the furthest time to plot

g = 25: arbitrary scaling factor for a generation

Further optional arguments are passed to plot from pyplot.

DemoPlots.plot_hist — Method

plot_hist(h::Histogram; kwargs...)

Plot the histogram h using PyPlot.

Optional arguments are passed to scatter from pyplot.

DemoPlots.plot_input — Method

plot_input(TN, ax; max_t = 5e6, g = 25, kwargs...)

Plot the demographic profile encoded in the parameters TN as input.

Arguments

max_t: the furthest time to plot

g = 25: arbitrary scaling factor for a generation

kwargs...: the keywords that PyPlot plot accepts

DemoPlots.plot_remnbps — Method

plot_remnbps(para::Vector, ax; max_t = 5e6, g = 25, kwargs...)

Plot the remaining number of base pairs as a function of time, given the parameters para.

ax is a pyplot axis.

Arguments

max_t = 5e6: the furthest time, in generations, at which the coalescent is evaluated

g = 25: arbitrary scaling factor for a generation

Further optional arguments are passed to plot and scatter from pyplot.

DemoPlots.plot_residuals_sim — Method

plot_residuals_sim(h_obs::Histogram, fit::DemoInfer.FitResult, μ::Float64, ρ::Float64; kwargs...)
plot_residuals_sim(h_obs::Histogram, para::Vector{T}, μ::Float64, ρ::Float64; kwargs...)

Plot the residuals of the simulation, with given fit resultorparaas input, with respect to the observed histogramh_obs`.

Arguments

factor = 1: the factor which determines how many times the simulation is repeated

Further optional arguments are passed to scatter from pyplot, ax is the pyplot axis.

DemoPlots.plot_residuals_th — Method

plot_residuals_th(h_obs::Histogram, fit::DemoInfer.FitResult, μ::Float64; kwargs...)
plot_residuals_th(h_obs::Histogram, para::Vector{T}, μ::Float64; kwargs...)

Plot of residuals between observed histogram h_obs and the theory.

Optional arguments are passed to scatter from pyplot, ax is the pyplot axis.

DemoPlots.xy — Method

xy(h::HistogramBinnings.Histogram{T, 1, E}; mode = :density) where {T, E}

Return the midpoints and the weights of the histogram h.

Arguments

h: the histogram

mode: the normalization mode, :density normalizes counts by bins widths